pkg/bpf: move maps memlock retrieval from bugtool to bpf
These operations might eventually be taken care of by cilium/ebpf
directly and we can remove this, see 21be59580a51 ("map: add map_extra,
memlock, frozen to MapInfo").

I moved this to the bpf package as it made little sense to import
pkg/bugtool to just perform those operations on maps.
