+++ pkg/bpf/maps_test.go
@@ -1,7 +1,7 @@
 // SPDX-License-Identifier: Apache-2.0
 // Copyright Authors of Tetragon
 
-package bugtool
+package bpf
 
 import (
 	"io"
